Refactoring - moved retry logic to utils file [ch11270]
