+++ lib/chartmogul/resource.js
@@ -1,8 +1,6 @@
 'use strict';
 
 const errors = require('./errors');
-const request = require('request');
-const retry = require('retry');
 const util = require('./util');
 const Config = require('./config');
 
@@ -67,7 +65,7 @@ class Resource {
         json: true
       };
 
-      retryRequest(config.retries, options, (error, response, body) => {
+      util.retryRequest(config.retries, options, (error, response, body) => {
         if (error) {
           return reject(error);
         }
@@ -153,54 +151,5 @@ class Resource {
 Object.keys(mappings).forEach(methodName => {
   Resource[methodName] = Resource._method(mappings[methodName]);
 });
-
-function retryOnStatus (res) {
-  if (
-    res &&
-    (res.statusCode === 429 || (res.statusCode >= 500 && res.statusCode < 600))
-  ) {
-    return new Error(`${res.statusCode} - ${res.statusMessage}`);
-  }
-}
-
-// ref: https://github.com/FGRibreau/node-request-retry/blob/master/strategies/NetworkError.js#L3:22
-var RETRIABLE_ERRORS = [
-  'ECONNRESET',
-  'ENOTFOUND',
-  'ESOCKETTIMEDOUT',
-  'ETIMEDOUT',
-  'ECONNREFUSED',
-  'EHOSTUNREACH',
-  'EPIPE',
-  'EAI_AGAIN'
-];
-
-function retryOnNetworkError (err) {
-  if (err && RETRIABLE_ERRORS.indexOf(err.code) > -1) {
-    return err;
-  }
-  return null;
-}
-
-function retryRequest (retries, options, cb) {
-  if (retries === undefined) {
-    retries = 20; // default is around 15min
-  }
-
-  var operation = retry.operation({
-    retries,
-    randomize: 0.5,
-    minTimeout: 500,
-    maxTimeout: 60 * 1000
-  });
-  operation.attempt(function (currentAttempt) {
-    request(options, (err, res, body) => {
-      if (operation.retry(retryOnStatus(res) || retryOnNetworkError(err))) {
-        return;
-      }
-      cb(err, res, body);
-    });
-  });
-}
 
 module.exports = Resource;
