+++ lib/chartmogul/util/index.js
@@ -12,3 +12,5 @@ exports.expandPath = function (path, data) {
 
   return template.fill(data);
 };
+
+exports.retryRequest = require('./retry');
